pyproject.toml

[project]
name = "panpipes"
version = "0.5.0"
authors = [
    {name = "Charlotte Rich-Griffin"},
    {name = "Fabiola Curion"}
]
description = "Panpipes - multimodal single cell pipelines"
readme = "README.md"

classifiers = [
    "License :: OSI Approved :: BSD License",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: R"
    ]

requires-python = ">= 3.10"

dependencies = [
    "jax",
    "jaxlib",
    "apsw",
    "bbknn",
    "cgatcore",
    "drmaa",
    "gevent",
    "harmonypy",
    "mofapy2",
    "muon",
    "matplotlib",
    "openpyxl",
    "paramiko",
    "pysam",
    "pyyaml",
    "ruffus",
    "scanorama",
    "scanpy", 
    "seaborn",
    "leidenalg",
    "louvain",
    "scib",
    "scib-metrics",
    "scikit-misc",
    "scirpy",
    "scrublet",
    "scvi-tools",
    "sqlalchemy",
    ]

[project.optional-dependencies]
dev = [
    "packaging",
    "pytest",
    "twine",
    "build"
]
notebook = [
    "jupyter",
    "jupyterlab"
]
spatial = [
    "jax",
    "jaxlib",
    "scvi-tools",
    "scipy",
    "squidpy",
    "cell2location",
    "tangram-sc", 
    "spatialdata", 
    "spatialdata-io", 
    "dask"
]

refmap_old = [
    "scvi-tools",
    "pandas<2.0.0",
    "scib"
]


[project.scripts]
panpipes = "panpipes:entry.main"

[tool.setuptools]
package-data = {"panpipes" = ["panpipes/*/*.yml", "R_scripts/*.R", "resources/*"]}


[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"
